make file structure consistent, fix bug
