+++ cypress/integration/pages/topicPage/index.js
@@ -1,6 +1,8 @@
 import config from '../../../support/config/services';
 import getPaths from '../../../support/helpers/getPaths';
 import serviceHasPageType from '../../../support/helpers/serviceHasPageType';
+import testsForCanonicalOnly from './testsForCanonicalOnly';
+import testsForAMPOnly from './testsForAMPOnly';
 import crossPlatformTests from './tests';
 import visitPage from '../../../support/helpers/visitPage';
 import { overrideRendererOnTest } from '../../../support/helpers/onDemandRadioTv';
@@ -27,6 +29,32 @@ Object.keys(config)
           pageType,
           variant,
         });
+        testsForCanonicalOnly({
+          service,
+          pageType,
+          variant,
+        });
       });
     });
+    paths
+      .map(path => `${path}.amp`)
+      .forEach(currentPath => {
+        describe(`${pageType} - ${currentPath}`, () => {
+          before(() => {
+            Cypress.env('currentPath', currentPath);
+            visitPage(currentPath, pageType);
+          });
+          crossPlatformTests({
+            service,
+            pageType,
+            variant,
+            isAmp: true,
+          });
+          testsForAMPOnly({
+            service,
+            pageType,
+            variant,
+          });
+        });
+      });
   });
